+++ R/plot.chisqtest.R
@@ -20,14 +20,12 @@
 #' @param signifdigitschisq A numeric indicating the number of desired significant figures reported for the \eqn{\chi^2}{\chi-squared} label (optional).
 #' @param curvelinesize A numeric indicating the size of the curve line (optional).
 #' @param cutlinesize A numeric indicating the size of the cut line (optional). By default, the size of the curve line is used.
+#' @param p_value_position A numeric vector of length 2, indicating the x and y coordinates of the p-value label. By default, the position is set to \code{"auto"}. Note that the absolute value is used, and the sign is ignored.
 #' @return A plot with the density of probability of \eqn{\chi^2}{\chi-squared} under the null hypothesis, annotated with the observed test statistic and the p-value.
 #' @export plotchisqtest
 #' @examples
 #' #Making a chi-squared plot with Chi-squared of 8 and df of 4
 #' plotchisqtest(chisq = 8, df = 4)
-#'
-#' #Note that the same can be obtained even quicker with:
-#' plotchisqtest(8,4)
 #'
 #' #The same plot without the Chi-Squared or p value
 #' plotchisqtest(8,4, blank = TRUE)
@@ -45,12 +43,12 @@
 #' plotchisqtest(comp)
 #'
 #' @author Nils Myszkowski <nmyszkowski@pace.edu>
-plotchisqtest <- function(chisq, df = chisq$parameter, blank = FALSE, xmax = "auto", title = parse(text = expression(chi^2 ~ "Test")), xlabel = parse(text = expression(chi^2)), ylabel = "Density of probability\nunder the null hypothesis", fontfamily = "serif", colorleft = "aliceblue", colorright = "firebrick3", colorleftcurve = "black", colorrightcurve = "black", colorcut = "black", colorplabel = colorright, theme = "default", signifdigitschisq = 3, curvelinesize = .4, cutlinesize = curvelinesize) {
+plotchisqtest <- function(chisq, df = chisq$parameter, blank = FALSE, xmax = "auto", title = parse(text = expression(chi^2 ~ "Test")), xlabel = parse(text = expression(chi^2)), ylabel = "Density of probability\nunder the null hypothesis", fontfamily = "serif", colorleft = "aliceblue", colorright = "firebrick3", colorleftcurve = "black", colorrightcurve = "black", colorcut = "black", colorplabel = colorright, theme = "default", signifdigitschisq = 3, curvelinesize = .4, cutlinesize = curvelinesize, p_value_position = "auto") {
   x=NULL
 
 
 
-  # If chisq is a t.test() object, then mine it to get t and df
+  # If chisq is a test() object, then mine it to get t and df
   if ("htest" %in% class(chisq)) {
     df <- chisq$parameter
     chisq <- chisq$statistic
@@ -99,10 +97,20 @@ plotchisqtest <- function(chisq, df = chisq$parameter, blank = FALSE, xmax = "au
   density <- function(x) stats::dchisq(x, df)
   #Use the maximum density (top of the curve) to use as maximum y axis value (start finding maximum at .2 to avoid very high densities values when the density function has a y axis asymptote)
   maxdensity <- stats::optimize(density, interval=c(0.2, xbound), maximum=TRUE)$objective
-  #To place the p value labels on the x axis, at the middle of the part of the curve they correspond to
-  x_plabel <- chisq+(xbound-chisq)/2
-  #Use the density corresponding to the given chisquare to place the label above (if this density is too high places the label lower in order to avoid the label being out above the plot)
-  y_plabel <- min(density(chisq)+maxdensity*.1, maxdensity*.7)
+
+
+  # Set the position of the p value automatically or manually
+  if (length(p_value_position) == 1) {
+    #To place the p value labels on the x axis, at the middle of the part of the curve they correspond to
+    x_plabel <- chisq+(xbound-chisq)/2
+    #Use the density corresponding to the given chisquare to place the label above (if this density is too high places the label lower in order to avoid the label being out above the plot)
+    y_plabel <- min(density(chisq)+maxdensity*.1, maxdensity*.7)
+  } else {
+    x_plabel <- abs(p_value_position[1])
+    y_plabel <- abs(p_value_position[2])
+  }
+
+
   #Define the fill color of the labels as white
   colorlabelfill <- "white"
   #Theme options
@@ -160,13 +168,15 @@ plotchisqtest <- function(chisq, df = chisq$parameter, blank = FALSE, xmax = "au
      #Cut line
      ggplot2::geom_vline(xintercept = chisq, colour = colorcut, linewidth = cutlinesize) +
      #p label
-     ggplot2::geom_label(ggplot2::aes(x_plabel,y_plabel,label = plab), fill = colorlabelfill, colour=colorplabel, family = fontfamily) +
+     #ggplot2::geom_label(ggplot2::aes(x_plabel,y_plabel,label = plab), fill = colorlabelfill, colour=colorplabel, family = fontfamily) +
+     ggplot2::annotate(geom = "label", x = x_plabel, y = y_plabel, label = plab, fill = colorlabelfill, colour = colorplabel, family = fontfamily) +
      #Chi squared label
-     ggplot2::geom_label(ggplot2::aes(chisq,maxdensity*(-.05),label = chisqlab), fill = colorlabelfill, colour=colorcut, family=fontfamily, parse = TRUE) +
+     #ggplot2::geom_label(ggplot2::aes(chisq,maxdensity*(-.05),label = chisqlab), fill = colorlabelfill, colour=colorcut, family=fontfamily, parse = TRUE) +
+     ggplot2::annotate(geom = "label", x = chisq, y = maxdensity*(-.05), label = chisqlab, fill = colorlabelfill, colour = colorcut, family = fontfamily, parse = TRUE) +
      #Add the title
      ggplot2::ggtitle(title) +
      #Axis labels
-     ggplot2::labs(x=xlabel,y=ylabel, size=10) +
+     ggplot2::labs(x=xlabel,y=ylabel, size = 10) +
      #Apply black and white ggplot theme to avoid grey background, etc.
      ggplot2::theme_bw() +
      #Remove gridlines and pass fontfamily argument to ggplot2